src/requirements.txt
# GCP SDKs
google-cloud-storage==2.16.0       # Blob storage (instead of azure-storage-blob)
google-auth==2.22.0                # Authentication (instead of azure-identity)
google-cloud-documentai==2.25.0    # OCR/Form Recognizer replacement
google-cloud-logging==3.10.0       # Structured logging (instead of ecs-logging)
google-cloud-vision==3.7.2


# API framework & web server
fastapi==0.109.1
uvicorn[standard]==0.22.0
gunicorn==20.1.0

# HTTP & requests
requests==2.31.0
urllib3==2.0.7
httpx==0.24.1

# Testing
pytest-cov==4.1.0
pytest-dotenv==0.5.2
parameterized==0.9.0

# FastAPI helpers
python-multipart==0.0.6
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.17.0
fastapi-health==0.4.0
python-decouple==3.8

# Computer vision / ML utils
opencv-contrib-python==4.7.0.72
pydantic<=2.0
networkx==2.5.1
shapely==2.0.1
tqdm==4.65.0
matplotlib==3.7.2

# Database / files
cloud-sql-python-connector[pymysql]==1.6.0
pdf2image
pypdfium2

